# cmd/server/openapi.yaml
openapi: 3.1.0
info:
  title: Echo API
  version: 1.0.0
  description: |
    The Echo feedback platform API.

    All request bodies are JSON; all responses are JSON. Authentication uses a
    Bearer JWT issued by `POST /v1/login` and carried in the `Authorization`
    header as `Bearer <access_token>`. Routes marked with a lock icon require
    this token; public routes do not.

    The dev server listens on `http://localhost:1323`.
servers:
  - url: http://localhost:1323
    description: Local development server
...
paths:
...
  /v1/verify-email:
    get:
      tags: [Auth]
      summary: Verify an email address
      description: |
        Validates the email-verification token supplied in the `token` query
        parameter and, on success, marks the matching employee's email as
        verified. Public and idempotent: verifying an already-verified email
        succeeds again with the same `200` response.
      operationId: verifyEmail
      security: []
      parameters:
        - $ref: "#/components/parameters/VerificationToken"
      responses:
        "200":
          description: Email verified.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required: [message]
              examples:
                ok:
                  value: { message: "email verified" }
        "400":
          description: Missing, malformed, expired, or wrong-purpose token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalid:
                  value: { message: "invalid or expired verification token" }
...
  /v1/employees:
    get:
      tags: [Employees]
      summary: List employees in the caller's organization
      description: |
        Returns one page of employees in the authenticated caller's
        organization, ordered by name ascending. An employee can only see
        their own organization's members. The response omits the `password`
        field. The caller is included in the list; the frontend should filter
        the caller out when populating a reviewee picker (the backend enforces
        no self-review at `POST /v1/feedbacks`).

        Pagination is cursor-based: pass the `next_cursor` value from one
        response as the `cursor` query parameter on the next request. When
        `next_cursor` is `null` there are no more pages. An unknown cursor
        (one that does not refer to an existing employee) returns `400`.
      operationId: listEmployees
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListCursor"
      responses:
        "200":
          description: One page of employees plus the next-page cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
        "400":
          description: Unknown cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                unknownCursor:
                  value: { message: "unknown cursor" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/employees/{id}/manager:
    patch:
      tags: [Employees]
      summary: Assign or clear an employee's manager
      description: |
        Sets (or, when `manager_id` is `null`, clears) the manager of the named
        employee. Only org admins may do this; the check is a fresh-load check
        server-side, so a stale token cannot grant privileges the caller no
        longer has. The target and the new manager must belong to the caller's
        organization.

        Validation: the new manager must exist, be in the same organization,
        not be the target themselves, and the assignment must not create a
        management cycle (the manager chain is walked server-side, capped at
        50 hops; a longer or already-cyclic chain is rejected). Clearing the
        manager (`manager_id: null`) skips these checks: removing a parent
        pointer cannot create a cycle.
      operationId: assignManager
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The employee ID whose manager is being assigned.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignManagerRequest"
            examples:
              assign:
                summary: Assign a manager
                value:
                  manager_id: 0190aaaa-0001-7000-8000-000000000000
              clear:
                summary: Clear the manager
                value:
                  manager_id: null
      responses:
        "200":
          description: The updated employee, with the new `manager_id`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeResponse"
        "400":
          description: |
            Malformed body or a failed validation (unknown manager ID,
            cross-org manager, self-assignment, or a would-be cycle).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                cycle:
                  value: { message: "assignment would create a management cycle" }
                unknownManager:
                  value: { message: "manager_id does not refer to an existing employee" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Caller is not an org admin, or the target is in another organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                notAdmin:
                  value: { message: "only org admins can assign managers within their organization" }
        "404":
          description: No employee matches the target ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                notFound:
                  value: { message: "employee not found" }
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/me/reports:
    get:
      tags: [Employees]
      summary: List the current employee's direct reportees
      description: |
        Returns one page of employees whose `manager_id` is the authenticated
        caller (i.e. the caller's direct reportees), ordered by name ascending.
        The manager is the authenticated employee (taken from the JWT subject),
        so an employee can only list their own reportees. An employee with no
        reportees gets an empty list. The response omits the `password` field.

        Pagination is cursor-based: pass the `next_cursor` value from one
        response as the `cursor` query parameter on the next request. When
        `next_cursor` is `null` there are no more pages. An unknown cursor
        (one that does not refer to an existing employee) returns `400`.
      operationId: listMyReports
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListCursor"
      responses:
        "200":
          description: One page of reportees plus the next-page cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
        "400":
          description: Unknown cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                unknownCursor:
                  value: { message: "unknown cursor" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    VerificationToken:
      name: token
      in: query
      required: true
      description: The email-verification JWT issued at registration.
      schema:
        type: string
      example: eyJhbGciOiJIUzI1NiIs...
    ListLimit:
      name: limit
      in: query
      required: false
      description: |
        Page size. Non-numeric or `<= 0` falls back to the default (20);
        values above `100` are capped.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    ListCursor:
      name: cursor
      in: query
      required: false
      description: |
        The `next_cursor` value from the previous page (an employee ID).
        Omit on the first page. An unknown cursor returns `400 "unknown cursor"`.
      schema:
        type: string
      example: 0190aaaa-0000-7000-8000-000000000000
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            unauthorized:
              value: { message: "missing or invalid token" }
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            internal:
              value: { message: "failed to register employee" }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A signed JWT obtained from `POST /v1/login`. Send as
        `Authorization: Bearer <access_token>`.
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: A human-readable description of the error.
      required: [message]
    Role:
      type: string
      description: The authorization role of an employee within their organization.
      enum:
        - org_admin
        - user
      examples:
        - org_admin
...
    EmployeeResponse:
      type: object
      description: The representation of an employee returned to clients. The `password` field is omitted.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        organization_name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        manager_id:
          type: string
          format: uuid
          nullable: true
          description: The employee ID of the user's manager, or null.
        title:
          type: string
        email:
          type: string
          format: email
        is_mail_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - organization_name
        - role
        - title
        - email
        - is_mail_verified
        - created_at
        - updated_at
      example:
        id: 0190aaaa-0000-7000-8000-000000000000
        name: Alice
        organization_name: Acme
        role: org_admin
        manager_id: null
        title: Senior Engineer
        email: alice@example.com
        is_mail_verified: true
        created_at: "2026-08-18T10:00:00Z"
        updated_at: "2026-08-18T10:00:00Z"
    EmployeeListResponse:
      type: object
      description: One page of employees plus the next-page cursor.
      properties:
        employees:
          type: array
          description: The employees on this page. Never null (empty array when no members).
          items:
            $ref: "#/components/schemas/EmployeeResponse"
        next_cursor:
          type: string
          format: uuid
          nullable: true
          description: |
            The ID of the last employee on this page; pass it as `cursor` on
            the next request. Null when there are no more pages.
      required: [employees, next_cursor]
      example:
        employees:
          - id: 0190aaaa-0000-7000-8000-000000000000
            name: Alice
            organization_name: Acme
            role: org_admin
            manager_id: null
            title: Senior Engineer
            email: alice@example.com
            is_mail_verified: true
            created_at: "2026-08-18T10:00:00Z"
            updated_at: "2026-08-18T10:00:00Z"
        next_cursor: 0190aaaa-0000-7000-8000-000000000000
...
    AssignManagerRequest:
      type: object
      description: Request body for `PATCH /v1/employees/{id}/manager`.
      properties:
        manager_id:
          type: string
          format: uuid
          nullable: true
          description: |
            The employee ID to assign as the target's manager, or `null` to
            clear the assignment. The manager must exist, belong to the same
            organization as the caller, not be the target themselves, and not
            create a management cycle; `null` skips these checks.
      required: [manager_id]
      example:
        manager_id: 0190aaaa-0001-7000-8000-000000000000
...